Rename `org.scalajs.testing.interface` to `...testing.bridge`.
Follow up to 2240515b60390591fba83be88fab67a31b25adf7.
